+++ lib/active_record/tenanted/database_tasks.rb
@@ -45,6 +45,17 @@ module ActiveRecord
         $stdout.puts "Dropped database '#{db_config.database}'" if verbose?
       end
 
+      def rollback_all(step = 1)
+        tenants.each do |tenant|
+          rollback_tenant(tenant, step: step)
+        end
+      end
+
+      def rollback_tenant(tenant = set_current_tenant, step: 1)
+        db_config = config.new_tenant_config(tenant)
+        rollback(db_config, step: step)
+      end
+
       def tenants
         config.tenants.presence || [ get_default_tenant ].compact
       end
@@ -115,12 +126,33 @@ module ActiveRecord
         end
       end
 
+      def rollback(config, step:)
+        ActiveRecord::Tasks::DatabaseTasks.with_temporary_connection(config) do |conn|
+          pool = conn.pool
+
+          pool.migration_context.rollback(step)
+          pool.schema_cache.clear!
+
+          if Rails.env.development? || ENV["ARTENANT_SCHEMA_DUMP"].present?
+            ActiveRecord::Tasks::DatabaseTasks.dump_schema(config)
+
+            cache_dump = ActiveRecord::Tasks::DatabaseTasks.cache_dump_filename(config)
+            ActiveRecord::Tasks::DatabaseTasks.dump_schema_cache(pool, cache_dump)
+          end
+        end
+      end
+
       def verbose?
         self.class.verbose?
       end
 
       def register_rake_tasks
         name = config.name
+
+        step_from_env = -> {
+          step = ENV["STEP"]
+          step.present? ? step.to_i : 1
+        }
 
         desc "Migrate tenanted #{name} databases for current environment"
         task "db:migrate:#{name}" => "load_config" do
@@ -138,6 +170,24 @@ module ActiveRecord
         end
         task "db:migrate" => "db:migrate:#{name}"
         task "db:prepare" => "db:migrate:#{name}"
+
+        desc "Rollback tenanted #{name} databases for current environment"
+        task "db:rollback:#{name}" => "load_config" do
+          verbose_was = ActiveRecord::Migration.verbose
+          ActiveRecord::Migration.verbose = ActiveRecord::Tenanted::DatabaseTasks.verbose?
+
+          step = step_from_env.call
+          tenant = ENV["ARTENANT"]
+
+          if tenant.present?
+            rollback_tenant(tenant, step: step)
+          else
+            rollback_all(step)
+          end
+        ensure
+          ActiveRecord::Migration.verbose = verbose_was
+        end
+        task "db:rollback" => "db:rollback:#{name}"
 
         desc "Drop tenanted #{name} databases for current environment"
         task "db:drop:#{name}" => "load_config" do